requirements.txt
Flask
Pandas
Matplotlib
mysql-connector-python
Flask-SQLAlchemy
fpdf2
openpyxl
flask-login
numpy
